+++ tests/TileIdTest.cpp
@@ -41,35 +41,35 @@ void TileIdTest::testFromCoordinates_data()
         int mid = qRound(last/2.0);
         int three = mid + quarter;
 
-        addRow() << -180.0 << 0.0 << zoom << 0 << mid;
-        addRow() <<  -90.0 << 0.0 << zoom << quarter << mid;
-        addRow() <<    0.0 << 0.0 << zoom << mid << mid;
-        addRow() <<   90.0 << 0.0 << zoom << three << mid;
-        addRow() <<  180.0 << 0.0 << zoom << last << mid;
+        addRow() << qreal(-180.0) << qreal(0.0) << zoom << 0 << mid;
+        addRow() <<  qreal(-90.0) << qreal(0.0) << zoom << quarter << mid;
+        addRow() <<    qreal(0.0) << qreal(0.0) << zoom << mid << mid;
+        addRow() <<   qreal(90.0) << qreal(0.0) << zoom << three << mid;
+        addRow() <<  qreal(180.0) << qreal(0.0) << zoom << last << mid;
 
-        addRow() << -180.0 << -89.9999 << zoom << 0 << last;
-        addRow() <<  -90.0 << -89.9999 << zoom << quarter << last;
-        addRow() <<    0.0 << -89.9999 << zoom << mid << last;
-        addRow() <<   90.0 << -89.9999 << zoom << three << last;
-        addRow() <<  180.0 << -89.9999 << zoom << last << last;
+        addRow() << qreal(-180.0) << qreal(-89.9999) << zoom << 0 << last;
+        addRow() <<  qreal(-90.0) << qreal(-89.9999) << zoom << quarter << last;
+        addRow() <<    qreal(0.0) << qreal(-89.9999) << zoom << mid << last;
+        addRow() <<   qreal(90.0) << qreal(-89.9999) << zoom << three << last;
+        addRow() <<  qreal(180.0) << qreal(-89.9999) << zoom << last << last;
 
-        addRow() << -180.0 << 89.9999 << zoom << 0 << 0;
-        addRow() <<  -90.0 << 89.9999 << zoom << quarter << 0;
-        addRow() <<    0.0 << 89.9999 << zoom << mid << 0;
-        addRow() <<   90.0 << 89.9999 << zoom << three << 0;
-        addRow() <<  180.0 << 89.9999 << zoom << last << 0;
+        addRow() << qreal(-180.0) << qreal(89.9999) << zoom << 0 << 0;
+        addRow() <<  qreal(-90.0) << qreal(89.9999) << zoom << quarter << 0;
+        addRow() <<    qreal(0.0) << qreal(89.9999) << zoom << mid << 0;
+        addRow() <<   qreal(90.0) << qreal(89.9999) << zoom << three << 0;
+        addRow() <<  qreal(180.0) << qreal(89.9999) << zoom << last << 0;
 
-        addRow() << -180.0 << -179.9999 << zoom << 0 << mid;
-        addRow() <<  -90.0 << -179.9999 << zoom << quarter << mid;
-        addRow() <<    0.0 << -179.9999 << zoom << mid << mid;
-        addRow() <<   90.0 << -179.9999 << zoom << three << mid;
-        addRow() <<  180.0 << -179.9999 << zoom << last << mid;
+        addRow() << qreal(-180.0) << qreal(-179.9999) << zoom << 0 << mid;
+        addRow() <<  qreal(-90.0) << qreal(-179.9999) << zoom << quarter << mid;
+        addRow() <<    qreal(0.0) << qreal(-179.9999) << zoom << mid << mid;
+        addRow() <<   qreal(90.0) << qreal(-179.9999) << zoom << three << mid;
+        addRow() <<  qreal(180.0) << qreal(-179.9999) << zoom << last << mid;
 
-        addRow() << -180.0 << 179.9999 << zoom << 0 << mid-1;
-        addRow() <<  -90.0 << 179.9999 << zoom << quarter << mid-1;
-        addRow() <<    0.0 << 179.9999 << zoom << mid << mid-1;
-        addRow() <<   90.0 << 179.9999 << zoom << three << mid-1;
-        addRow() <<  180.0 << 179.9999 << zoom << last << mid-1;
+        addRow() << qreal(-180.0) << qreal(179.9999) << zoom << 0 << mid-1;
+        addRow() <<  qreal(-90.0) << qreal(179.9999) << zoom << quarter << mid-1;
+        addRow() <<    qreal(0.0) << qreal(179.9999) << zoom << mid << mid-1;
+        addRow() <<   qreal(90.0) << qreal(179.9999) << zoom << three << mid-1;
+        addRow() <<  qreal(180.0) << qreal(179.9999) << zoom << last << mid-1;
     }
 }
